+++ src/Copse.Async/Treenumerators/AsyncLevelOrderStreamBreadthFirstTreenumerator.cs
@@ -1,3 +1,4 @@
+using Copse.Collections;
 using Copse.Stores;
 using Copse.Core;
 using System.Runtime.CompilerServices;