+++ src/Copse.Linq.Async/TreenumerableExtensions/Treenumerable/AsyncTreenumerable.OrderChildrenBy.cs
@@ -1,3 +1,4 @@
+using Copse.Collections;
 using Copse;
 using Copse.Stores;
 using Copse.Treenumerables;