+++ src/Copse.Linq.Async/Treenumerables/Buffers/AsyncLevelOrderAdjacencyIndex.cs
@@ -1,3 +1,4 @@
+using Copse.Collections;
 using Copse.Core;
 using Copse;
 using Copse.Stores;