+++ src/Copse.Linq.Async/Treenumerators/SelectMany/AsyncSelectManyDepthFirstTreenumerator.cs
@@ -1,3 +1,4 @@
+using Copse.Collections;
 using Copse;
 using Copse.Core;
 using System;