+++ src/Copse/Traversal/BreadthFirstPathState.cs
@@ -1,3 +1,4 @@
+using Copse.Collections;
 using Copse.Core;
 using System;
 using System.Runtime.CompilerServices;